Remove RegexMatcherRule field filtre_document
This field contains a regexp that was executed on the raw text in order
to determine if the rule should be applied to the document.
This is a very specific use case so not worth keep. The proper way to do
this would be to create another Annotator identifying entities on which
the rules should be applied, then instantiate another RegexpMatcher with
these rules and run in on the identified entities.