Add unicode_sensitive entry in NegationDetectorRule
Lots of rules in negation_detector_default_rules.yml are trying to
do this by regrouping all variants of a character in a character class
(ex [eéè]) but they are not doing this consistently and this is why
some tests fail.

Using unidecode.unidecode seems like the right way to do it (this is
used by quickumls when unicode_normalize=True). It is still possible to
create "unicode-sensitive" rules if necessary.

Not that unless unicode_sensitive is True, rules shouldn't contain
non-ASCII char since they will never be matched.