Fix end-of-word boundary in negation_detector_default_rules.yml
Usage of [^a-z] in negation_detector_default_rules.yml is always at the
beginning or end of a regexp, ex: '(^|[^a-z])infirmant\s*pas[^a-z]'.
Here the idea is to discriminate from a longer word beginning or ending
with "pas" (or some other word in other cases)

At the end of the regexp we must use ($|[^a-z]) because otherwise we
won't match strings ending directly with "pas". This is why some tests
are failing.