Rewrite matching optional intermediate words pattern in negation_detector_default_rules.yml
In negation_detector_default_rules.yml, a recurring pattern is to allow
1 or 2 optional intermediate words between 2 words, for instance:

    (^|[^a-z])pas\s*([a-z]*\s){0,2}doute

The problem with the way the regex is currently written is that the
space between the 2 outer words is optional ("pasdoute" will match),
as well as the same between the last inner word and "doute"
("pas de redoute" will match). Also intermediate words must be separated
by exactly one space. So this is rewritten to:

    (^|[^a-z])pas\s+([a-z]+\s+){0,2}doute

Sometimes apostrophes are also allowed, such as in :

    (^|[^a-z])pas\s*([a-z'']*\s*){0,2}elimine

The intent here is probably to allow cases such as "pas d'éliminer".
This has the same limits has the previous regex, so this is replaced by:

    (^|[^a-z])pas\s+([a-z]+['\s]+){0,2}elimine

Here ' is treated as a word separator, like \s. The only caveat is this
allows multiple consecutive '.

Note that these pattern still aren't very robust. Compounds are not
supported for instance, and it is not very clear why sometimes
apostrophes are allowed in intermediate words and other times not.