Make exclusion regexs exhaustive in negation_detector_default_rules.yml
Sometimes exclusion_regexps have to be repeated across several rules.
For instance, if we want "(^|[^a-z])sans\s+infirme" to not be matched,
we have to add it to the exclusions regexp of both the "sans" rule,
(which matches anything containing "sans") and of the
"infirme" rule (which matches anything containing "infirme").

Another way to handle this could be to always test all exclusion_regexps
from all rules but it seems like a bad idea. It would really differ
from the semantic implied by the format of the rules list, and it would
make it harder to understand why something was matched or not. It is
better to keep things scoped. Also this is closer to how the
RegexpMatcher works.