Make processing operations work on a list of annotations
All annotators now directly receive the list of input annotations they
have to process, they are not in charge of retrieving them on the
document/collection.

They can retrieve one or several list of created annotations.

The pipeline will be in charge of feeding each annotator with the
input annotations they needed, and of handling the annotations they
return.