Use new ProvBuilder in all annotators
Each annotator is now able to be set an optional ProvBuilder instance
to which it will provide provenance info (the data item created, the
operation that created it, the source data items it was derived from).

If the ProvBuilder is not set, it is still possible to use the annotator
but the provenance info will be lost.

It is also possible to create annotators that cannot be set a
ProvBuilder and therefore do not support the tracking of provenance for
annotation they created. But they might "break the provenance chain" if
used with other annotators that do support provenance.

In the next commit the "origin" field in annotations will be dropped.