Split Pipeline into a DocPipeline and a Pipeline operation
DocPipeline has the same API as the Pipeline class before but internally
use the new Pipeline class.

The new Pipeline is now an ProcessingOperation, dealing with list of
annotations as inputs and returning newly created annotations.
This makes it possible to nest pipelines (ie use a Pipeline as a step
of another Pipeline).

For now there is still a mechanism to provide a document to the Pipeline
(set_doc()), because the Pipeline needs to know the document in order
to append to it all generated intermediary annotations that are not
returned.