Introduce ProvBuilder and ProvGraph
ProvGraph contains list of ProvNodes. Each ProvNode is a vertex
representing an element of data (aka "data item"), typically an
annotation). The ProvNode also contains the id of the operation that
generated the data item, and the list of the source data items it was
derived from. This list represents the edges connected to the vertex.

The ProvBuilder is a class that helps to progressively build the
ProvGraph. Each operation will use it to add information about the data
items it generated and how they were generated. It doesn't do much for
now but it will handle more things in upcoming commits.