Fix type hinting issues with Document/Annotation
Before this commit, the way we subclassed Document and Annotation was
not valid according to mypy because it violated the liskov substitution
principle: TextDocument.add_annotation() could not be passed any
instance of Annotation, as defined in base class.

This is solved using Generic with an associated AnnotationType type,
that is bound to be an Annotation or an Annotation subclass.