Store intermediate annotations and operation descriptions in new ProvStore instead of Document
The ProvStore is responsible of storing all data items and operation
descriptions referenced by a ProvGraph, instead storing them in a
Document.

Among other things, this will be more convenient when handling
provenance info about annotations attached to different Documents but
referenced in a common ProvGraph (because there is no way to know in
which Document we should retrieve the annotation otherwise).

This way it is also easier to distinguish between annotations that
are of interested and must be attached back to a Document and
intermediate annotations that are only needed for provenance.

After this commit, all annotations and operations are not systematically
stored on the Document anymore. But the DocPipeline still adds
"output annotations" (but not the intermediate annotations) to each doc.