In HFTranslator, simplify handling of None tokens
Before this commit, we excluded the first and last tokens when
retrieving the embeddings. Because of this we had 2 different sets of
token indices referring to the same tokens, which was confusing.

Also, the new way we are now handling this is now able to handle padded
encoding (with several trailing Nones).