[regexp-matcher] Stop passing flags to Pattern.match and Pattern.finditer
The re.IGNORECASE flag is passed when initializing the pattern. The
second argument of Pattern.match() or Pattern.finditer() is not a flag
but a start position. No error was raised because the value of
re.IGNORECASE is 2.

Before this fix it was not possible to match an entity starting at the
1st or 2d character of a segment for case-insensitive rules.